+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 - Default style for `print-table` can be overridden via a dynamic var
 - Use a vertical bar (│) not a pipe character (|) in binary output
-
+- New `clj-commons.pretty.nrepl` namespace to setup pretty inside nREPL
 
 ## 3.4.1 -- 23 Jun 2025